requirements.txt
pyyaml>=6.0
elasticsearch>=8.0.0
